feat: update preview styling
